Fix review items: dbt materialization config, test fixture type, README consistency
- dbt_project.yml: intermediate materialization view → table (matches model-level config)
- conftest.py: num_holdings INTEGER → BIGINT (matches actual pipeline schema)
- sqlmesh README: uv sync → just sync (consistent with other READMEs)

+++ packages/mcp/tests/conftest.py
@@ -94,7 +94,7 @@ def test_db(tmp_path_factory):
     conn.execute("""
         CREATE TABLE marts.fct_portfolio_waci (
             portfolio_id VARCHAR, as_of_date DATE, waci DOUBLE,
-            total_market_value DOUBLE, num_holdings INTEGER
+            total_market_value DOUBLE, num_holdings BIGINT
         )
     """)
     conn.execute("""
